ENH inline image_benchmark example using ExampleBenchmark strings
Remove the directory-based image_benchmark and rewrite
run_image_benchmark.py to define all components (objective, dataset,
solvers, plot) as inline strings, following the same pattern as
run_benchopt_prepare.py.
